requirements.txt


numpy
matplotlib
Pillow
tensorflow==1.14

